Clean-up and change ints to doubles for database

+++ src/main/resources/queries/create_tables.sql
@@ -1,5 +1,5 @@
 -- Created by Vertabelo (http://vertabelo.com)
--- Last modification date: 2018-01-20 03:01:39.154
+-- Last modification date: 2018-01-20 15:12:41.906
 
 -- tables
 -- Table: home
@@ -8,11 +8,11 @@ USE winhomes;
 CREATE TABLE home (
     UUID varchar(36) NOT NULL,
     Name varchar(64) NOT NULL,
-    X int NOT NULL,
-    Y int NOT NULL,
-    Z int NOT NULL,
-    yaw int NOT NULL,
-    pitch int NOT NULL,
+    X double(12,4) NOT NULL,
+    Y double(12,4) NOT NULL,
+    Z double(12,4) NOT NULL,
+    yaw double(12,4) NOT NULL,
+    pitch double(12,4) NOT NULL,
     world varchar(64) NOT NULL,
     CONSTRAINT home_pk PRIMARY KEY (UUID)
 );
@@ -27,7 +27,8 @@ CREATE TABLE home_invite (
 -- foreign keys
 -- Reference: home_invite_home (table: home_invite)
 ALTER TABLE home_invite ADD CONSTRAINT home_invite_home FOREIGN KEY home_invite_home (home_UUID)
-    REFERENCES home (UUID);
+REFERENCES home (UUID);
 
 -- End of file.
+
